authentication template

+++ app/admin/admin.py
@@ -1,4 +1,4 @@
-from flask import Blueprint, render_template, flash
+from flask import Blueprint, render_template, flash, session, redirect, url_for
 from app.models import Users
 from app import db
 
@@ -7,16 +7,35 @@ admin = Blueprint("admin", __name__, static_folder="static", template_folder="te
 @admin.route("/view")
 @admin.route("/")
 def view():
-    return render_template("admin/view.html", values=Users.query.all())
+    if "email" in session:
+        email = session["email"]
+        if email == admin:
+            return redirect(url_for("admin.view"))
+        else:
+            flash("You're not an admin", "info")
+            return redirect(url_for("auth.login"))
+    else:
+        flash("You're not logged in", "info")
+        return redirect(url_for("auth.login"))
+
 
 @admin.route("/delete/<name>")
 def delete(name):
-    # db
-    found_user = Users.query.filter_by(name=name)
-    if found_user.delete():
-        db.session.commit()
-        flash("This user was deleted!", "info")
-    else:
-        flash("This name isn't in db!", "info")
+    if "email" in session:
+        email = session["email"]
+        if email == admin:
+            # db
+            found_user = Users.query.filter_by(name=name)
+            if found_user.delete():
+                db.session.commit()
+                flash("This user was deleted!", "info")
+            else:
+                flash("This name isn't in db!", "info")
 
-    return render_template("admin/view.html", values=Users.query.all())
+            return render_template("admin/view.html", values=Users.query.all())
+        else:
+            flash("You're not an admin", "info")
+            return redirect(url_for("auth.login"))
+    else:
+        flash("You're not logged in", "info")
+        return redirect(url_for("auth.login"))
